refactor: extract SIR optimization policy
