+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "crates/celox-design",
     "crates/celox-frontend-veryl",
     "crates/celox-sir",
+    "crates/celox-sir-opt",
     "crates/celox-slt",
     "crates/celox-state-layout",
     "crates/celox-testbench",
