+++ crates/celox/Cargo.toml
@@ -18,6 +18,7 @@ celox-analysis        = { path = "../celox-analysis" }
 celox-design          = { path = "../celox-design" }
 celox-frontend-veryl  = { path = "../celox-frontend-veryl" }
 celox-sir             = { path = "../celox-sir" }
+celox-sir-opt         = { path = "../celox-sir-opt" }
 celox-slt             = { path = "../celox-slt" }
 celox-state-layout    = { path = "../celox-state-layout" }
 celox-testbench       = { path = "../celox-testbench" }
